Update game.py
I added a box called rect3 on the game to keep track of the times clicked. 

I also went through and relabeled some things rect1 and rect2 so that we can work with each rectangle individually instead of having one set rectangle width and height. 

Just so you guys know, the text for the times clicked is rendered inside the while loop so that it is refreshed with the updated times_clicked value.