Update game.py
At the moment I have not found a way to implement one function to render the circle so we can calculate with it and one to draw the circle. 
This leads to some strange code in the running method of the Game class where the circle is drawn so we can calculate with it and then drawn once again later after we fill the screen black.
My current workaround for this is to draw the first circle with color values [0, 0, 0] and draw the second with the correct color values.
I think there is a way to simplify this but I have made lots of changes to the code and I want you guys to look at them before we proceed.